New TS muxer and associated files
This is a new, rewritten from scratch, TS muxer that aims at T-STD
compliance. It is licensed under the LGPL. It is a stream out module
instead of a mux module because it allows a finer control. Therefore the
syntax is not the same as the old TS mux:
vlc rtp://@:1234 --sout="#transcode:ts:udp"

It works better if incoming streams are CBR. It will probably not work
currently when muxing from file to file (async mode).